refactoring of models.py
