--- backend/ai_engine/app/question_gen/schemas.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-from typing import List, Dict
-
-class VideoResponse(BaseModel):
-    segments: List[Dict]
-    questions: List[Dict]
